+++ src/modules/marketplace-products/pages/marketplace-products-edit.vue
@@ -11,7 +11,32 @@
     @collapse="$emit('collapse:blade')"
   >
     <template #actions>
-      <mp-product-status :status="(product as ISellerProduct).status"></mp-product-status>
+      <div class="tw-flex tw-flex-row tw-items-center">
+        <div class="vc-status">
+          <VcSelect
+            name="currentLocale"
+            :model-value="currentLocale"
+            :options="localesOptions"
+            option-value="value"
+            option-label="label"
+            :disabled="disabled"
+            required
+            :clearable="false"
+            @update:model-value="
+              (e: string) => {
+                setLocale(e);
+              }
+            "
+          >
+          </VcSelect>
+        </div>
+        <div
+          v-if="(product as ISellerProduct).status !== 'Published'"
+          class="tw-ml-4"
+        >
+          <mp-product-status :status="(product as ISellerProduct).status"></mp-product-status>
+        </div>
+      </div>
     </template>
 
     <!-- Blade contents -->
@@ -148,27 +173,41 @@
                       @update:model-value="handleChange"
                     ></VcInput>
                   </Field>
-                  <Field
-                    v-slot="{ field, errorMessage, handleChange }"
-                    :label="$t('MP_PRODUCTS.PAGES.DETAILS.FIELDS.DESCRIPTION.TITLE')"
-                    name="description"
-                    rules="min:3|required"
-                    :model-value="productDetails.description"
+                  <div
+                    v-for="(lang, index) in localesOptions"
+                    :key="index"
                   >
-                    <VcEditor
-                      v-bind="field"
-                      v-model="productDetails.description"
-                      class="tw-mb-4"
+                    <Field
+                      v-if="lang.value == currentLocale"
+                      v-slot="{ field, errorMessage, handleChange }"
                       :label="$t('MP_PRODUCTS.PAGES.DETAILS.FIELDS.DESCRIPTION.TITLE')"
-                      :placeholder="$t('MP_PRODUCTS.PAGES.DETAILS.FIELDS.DESCRIPTION.PLACEHOLDER')"
-                      :disabled="disabled"
                       name="description"
-                      required
-                      :error-message="errorMessage"
-                      :assets-folder="productData.id || productData.categoryId"
-                      @update:model-value="handleChange"
-                    ></VcEditor>
-                  </Field>
+                      rules="min:3|required"
+                      :model-value="productDetails.descriptions.find((x) => x.languageCode == currentLocale)"
+                    >
+                      <VcEditor
+                        v-bind="field"
+                        v-model="productDetails.descriptions.find((x) => x.languageCode == currentLocale).content"
+                        class="tw-mb-4"
+                        :label="$t('MP_PRODUCTS.PAGES.DETAILS.FIELDS.DESCRIPTION.TITLE')"
+                        :placeholder="$t('MP_PRODUCTS.PAGES.DETAILS.FIELDS.DESCRIPTION.PLACEHOLDER')"
+                        :disabled="disabled"
+                        name="description"
+                        required
+                        :error-message="errorMessage"
+                        :assets-folder="productData.id || productData.categoryId"
+                        :languages="languages"
+                        :current-language="currentLocale"
+                        :multilanguage="true"
+                        @update:model-value="handleChange"
+                        @update:current-language="
+                        (e: string) => {
+                          setLocale(e);
+                        }
+                      "
+                      ></VcEditor>
+                    </Field>
+                  </div>
 
                   <VcDynamicProperty
                     v-for="property in filteredProps"
@@ -274,6 +313,7 @@ import {
   Property,
   PropertyValue,
   PropertyDictionaryItem,
+  EditorialReview,
 } from "../../../api_client/marketplacevendor";
 import { useIsFormValid, Field, useForm } from "vee-validate";
 import { min, required } from "@vee-validate/rules";
@@ -319,6 +359,7 @@ const {
   fetchCategories,
   revertStagedChanges,
   searchDictionaryItems,
+  getLanguages,
 } = useProduct();
 
 const { searchOffers } = useOffers();
@@ -356,6 +397,14 @@ const validateGtin = [
   async (value: string): Promise<string | boolean> => await validate("gtin", value),
 ];
 
+let languages: string[];
+let localesOptions;
+const currentLocale = ref("en-US");
+
+const setLocale = (locale: string) => {
+  currentLocale.value = locale;
+};
+
 const validate = _.debounce(
   async (fieldName: string, value: string): Promise<string | boolean> => {
     const sellerProduct = {
@@ -381,6 +430,9 @@ const validate = _.debounce(
 );
 
 const reload = async (fullReload: boolean) => {
+  languages = await getLanguages();
+  localesOptions = languages.map((x) => ({ label: t(`MP_PRODUCTS.PAGES.DETAILS.LANGUAGES.${x}`), value: x }));
+
   if (!modified.value && fullReload) {
     try {
       productLoading.value = true;
@@ -400,6 +452,17 @@ const reload = async (fullReload: boolean) => {
         sellerProductId: props.param,
       })
     )?.totalCount;
+  } else {
+    productDetails.value.descriptions = productDetails.value.descriptions.concat(
+      languages.map(
+        (x) =>
+          new EditorialReview({
+            languageCode: x,
+            content: "",
+            reviewType: "QuickReview",
+          })
+      )
+    );
   }
 };